+++ src/semantic-anlyzr/CMakeLists.txt
@@ -3,6 +3,9 @@ cmake_minimum_required(VERSION 3.16)
 add_executable(palan-sa
 	main.cpp
 	PlnSemanticAnalyzer.cpp
+	PlnSaExpr.cpp
+	PlnSaDecl.cpp
+	PlnSaStmt.cpp
 	PlnSaMessage.cpp
 	PlnType.cpp
 )
